refactor(core): cleanup & restructure shared code for hydra-master-stream10
