+++ src/main/java/faang/school/postservice/PostServiceApp.java
@@ -11,7 +11,6 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 @EnableRetry
 @EnableScheduling
 @EnableFeignClients(basePackages = "faang.school.postservice.client")
-@EnableRetry
 public class PostServiceApp {
     public static void main(String[] args) {
         new SpringApplicationBuilder(PostServiceApp.class)
