+++ src/main/java/faang/school/postservice/controller/LikeController.java
@@ -1,20 +1,31 @@
 package faang.school.postservice.controller;
 
+import faang.school.postservice.dto.LikeDto;
 import faang.school.postservice.dto.user.UserDto;
+import faang.school.postservice.exception.DataValidationException;
+import faang.school.postservice.service.LikeService;
 import faang.school.postservice.service.LikeServiceInterface;
 import jakarta.validation.constraints.Min;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-@RequiredArgsConstructor
+import java.util.Objects;
+
+
 @RestController
 @RequestMapping("/likes")
+@Slf4j
+@RequiredArgsConstructor
 public class LikeController {
     private final LikeService likeService;
     private final LikeServiceInterface likeServiceInterface;
