--- src/main/java/faang/school/postservice/dto/like/LikeDto.java
@@ -1,15 +1,0 @@
-package faang.school.postservice.dto.like;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class LikeDto {
-    private Long id;
-    private Long userId;
-    private Long commentId;
-    private Long postId;
-}
